feat: upgrade next js to 15.11 + react to 19
