--- app/dashboard/features/Todos/types.ts
@@ -1,6 +1,0 @@
-export interface Todo {
-  title: string;
-  body: string;
-  done: boolean;
-  id: number;
-}
