+++ app/features/Auth/LoginModal/types.ts
@@ -1,0 +1,9 @@
+export enum SignUpType {
+    login = "login",
+    register = "register",
+  }
+  
+ export enum AuthType {
+    biometry = "biometry",
+    default = "default",
+  }
