+++ app/hocs/interceptorHoc.tsx
@@ -1,7 +1,7 @@
 "use client";
 import React, { ReactNode } from "react";
-import { useAxiosInterceptor as useChatInterceptor } from "../dashboard/utils/chatFetcher";
-import { useAxiosInterceptor } from "../dashboard/utils/fetcher";
+import { useAxiosInterceptor as useChatInterceptor } from "../utils/chatFetcher";
+import { useAxiosInterceptor } from "../utils/fetcher";
 
 const InterceptorHoc: React.FC<{ children: ReactNode }> = ({ children }) => {
   // Add your logic or interceptors here
