+++ app/hooks/useAuthClient.ts
