+++ app/layout.tsx
@@ -1,14 +1,13 @@
 import "@styles/globals.css";
 import type { Metadata } from "next";
 
-import { AntdRegistry } from "@ant-design/nextjs-registry";
 import { cookies } from "next/headers";
-import UserProviderWrapper from "./dashboard/provider/userProvider";
-import { useAxiosSSRInterceptor } from "./dashboard/utils/serverSideFetcher";
 import InterceptorHoc from "./hocs/interceptorHoc";
 import AppLayout from "./layer";
+import UserProviderWrapper from "./providers/userProvider";
 
 import { Inter } from "next/font/google";
+import { subscribeSSRInterceptor } from "utils/serverSideFetcher";
 
 // If loading a variable font, you don't need to specify the font weight
 const inter = Inter({ subsets: ["latin"] });
@@ -18,8 +17,8 @@ export const metadata: Metadata = {
   description: "Generated by create next app",
 };
 
-async function useUseAccessToken() {
-  const accessToken = cookies().get("access_token")?.value;
+async function getAccessTokenFromCookies() {
+  const accessToken = (await cookies()).get("access_token")?.value;
 
   return { accessToken: accessToken ?? "" };
 }
@@ -29,19 +28,17 @@ export default async function RootLayout({
 }: Readonly<{
   children: React.ReactNode;
 }>) {
-  useAxiosSSRInterceptor();
-  const { accessToken } = await useUseAccessToken();
+  subscribeSSRInterceptor();
+  const { accessToken } = await getAccessTokenFromCookies();
 
   return (
     <html lang="en">
       <body className={inter.className} style={{ margin: 0 }}>
-        <AntdRegistry>
-          <UserProviderWrapper accessToken={accessToken}>
-            <InterceptorHoc>
-              <AppLayout>{children}</AppLayout>
-            </InterceptorHoc>
-          </UserProviderWrapper>
-        </AntdRegistry>
+        <UserProviderWrapper accessToken={accessToken}>
+          <InterceptorHoc>
+            <AppLayout>{children}</AppLayout>
+          </InterceptorHoc>
+        </UserProviderWrapper>
       </body>
     </html>
   );
