+++ package.json
@@ -10,15 +10,13 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@ant-design/icons": "^5.4.0",
-    "@ant-design/nextjs-registry": "^1.0.1",
     "@bufbuild/connect-web": "^0.13.0",
     "@bufbuild/protobuf": "^1.10.0",
     "@bufbuild/protoc-gen-connect-es": "^0.13.0",
     "@connectrpc/connect": "^1.5.0",
     "@connectrpc/connect-web": "^1.5.0",
     "@heroicons/react": "^2.2.0",
-    "@next/env": "^14.2.12",
+    "@next/env": "15.1.2",
     "@primer/octicons-react": "^19.11.0",
     "@redux-devtools/extension": "^3.3.0",
     "@reduxjs/toolkit": "^2.2.7",
@@ -28,10 +26,10 @@
     "axios": "^1.7.7",
     "dayjs": "^1.11.13",
     "jwt-decode": "^4.0.0",
-    "next": "14.2.11",
+    "next": "15.1.2",
     "next-themes": "^0.4.3",
-    "react": "^18",
-    "react-dom": "^18",
+    "react": "19.0.0",
+    "react-dom": "19.0.0",
     "react-icons": "^5.3.0",
     "react-redux": "^9.1.2",
     "redux": "^5.0.1",
@@ -43,15 +41,19 @@
     "@bufbuild/protoc-gen-es": "^1.10.0",
     "@connectrpc/protoc-gen-connect-es": "^1.5.0",
     "@types/node": "^20",
-    "@types/react": "^18",
-    "@types/react-dom": "^18",
+    "@types/react": "19.0.2",
+    "@types/react-dom": "19.0.2",
     "eslint": "^8",
-    "eslint-config-next": "14.2.11",
+    "eslint-config-next": "15.1.2",
     "husky": "^9.1.6",
     "postcss": "^8.4.45",
     "postcss-simple-vars": "^7.0.1",
     "shell-quote": "^1.8.1",
     "tailwindcss": "^3.4.1",
     "typescript": "^5"
+  },
+  "overrides": {
+    "@types/react": "19.0.2",
+    "@types/react-dom": "19.0.2"
   }
 }
